% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conforms_to_delta_fileorg.R
\name{conforms_to_delta_fileorg}
\alias{conforms_to_delta_fileorg}
\title{Checks if a directory for a delta value conforms to the expected output}
\usage{
conforms_to_delta_fileorg(delta_dir)
}
\arguments{
\item{delta_dir}{The delta directory.}
}
\value{
A boolean value for wether the directory conforms to the expected file organization.
}
\description{
Checks if a directory for a delta value conforms to the expected output
}
