% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hn2_subnetworks.R
\name{get_hn2_subnetworks}
\alias{get_hn2_subnetworks}
\title{Read in HotNet2 results}
\usage{
get_hn2_subnetworks(output_dir)
}
\arguments{
\item{output_dir}{The output directory from HotNet2.}
}
\value{
An S3 object with the parsed results from HotNet2.
}
\description{
Point this function at the output directory from HotNet2 and it will retrieve
the results from that analysis.
}
