% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_k_for_delta.R
\name{get_k_for_delta}
\alias{get_k_for_delta}
\title{Read in the "k" values for a delta directory}
\usage{
get_k_for_delta(delta_dir)
}
\arguments{
\item{delta_dir}{The direcotry for a delta value.}
}
\value{
A tibble with columns "size", "expected", "actual", and "pval".
}
\description{
Read in the "k" values for a delta directory
}
